+++ R/IPA_fr.R
@@ -23,7 +23,7 @@
 }
 
 ipa_fr <- function(word = "comportamento") {
-  fr_ipa_lex <- .get_user_lex("fr_ipa_lex")
+  fr_ipa_lex <- .lex("fr_ipa_lex")
 
   wd <- stringr::str_to_lower(word) |>
     stringr::str_remove_all("[:punct:]")