+++ R/IPA_pt_vec.R
@@ -9,8 +9,8 @@
 #' @noRd
 
 ipa_pt_vec <- function(word = c("palavra"), narrow = FALSE) {
-  pt_ipa_lex <- .get_user_lex("pt_ipa_lex")
-  pt_lex_user <- .get_user_lex("pt_lex_user")
+  pt_ipa_lex <- .lex("pt_ipa_lex")
+  pt_lex_user <- .lex("pt_lex_user")
 
   wd <- stringr::str_to_lower(word) |>
     stringr::str_remove_all(pattern = "[:punct:]") |>